uint8_t to uint16_t

+++ include/pipelineReg.h
@@ -7,7 +7,7 @@ using namespace std;
 // Slide p.85
 typedef struct control_EX {
     bool RegDst;
-    uint8_t ALUOp;
+    uint16_t ALUOp;
     bool ALUSrc;
 } control_EX;
 
@@ -24,9 +24,9 @@ typedef struct control_WB {
 
 typedef struct IF_ID_Pipeline_Reg {
     string OpCode; // "lw", "sub", "add", "beq", "sw";
-    uint8_t RegRs;
-    uint8_t RegRt;
-    uint8_t RegRd;
+    uint16_t RegRs;
+    uint16_t RegRt;
+    uint16_t RegRd;
     int16_t Immediate;
 } IF_ID_Pipeline_Reg;
 
@@ -37,9 +37,9 @@ typedef struct ID_EX_Pipeline_Reg {
     int32_t ReadData1;
     int32_t ReadData2;
     int32_t Immediate;
-    uint8_t RegRs;
-    uint8_t RegRt;
-    uint8_t RegRd;
+    uint16_t RegRs;
+    uint16_t RegRt;
+    uint16_t RegRd;
 } ID_EX_Pipeline_Reg;
 
 typedef struct EX_MEM_Pipeline_Reg {
@@ -48,14 +48,14 @@ typedef struct EX_MEM_Pipeline_Reg {
     bool Zero;
     int32_t ALU_Result;
     int32_t ReadData;
-    uint8_t RegRd;
+    uint16_t RegRd;
 } EX_MEM_Pipeline_Reg;
 
 typedef struct MEM_WB_Pipeline_Reg {
     control_WB Ctl_WB;
     int32_t DataOfMem;
     int32_t ALU_Result;
-    uint8_t RegRd;
+    uint16_t RegRd;
 } MEM_WB_Pipeline_Reg;
 
 extern IF_ID_Pipeline_Reg IF_ID_Reg;
